Extract/requirements.txt
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
click>=8.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
